% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDistanceToPerformanceScenario.R
\name{calcDistanceToPerformanceScenario}
\alias{calcDistanceToPerformanceScenario}
\title{Attach portfolio performance and distance to target}
\usage{
calcDistanceToPerformanceScenario(x)
}
\arguments{
\item{x}{An optimized optimLanduse object.}
}
\value{
An optimized optimLanduse object with attached portfolio performance.
}
\description{
The function calculates and attaches the portfolio performance and distance to target. See Gosling, Reith, Knoke, Gerique and Paul (2020): Exploring farmer perceptions of agroforestry via multi-objective optimisation: a test application in Eastern Panama. Formula S5 (supplementary).
}
