% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreperation.R
\name{dataPreperation}
\alias{dataPreperation}
\title{Preperation of the data for the robust optimization}
\usage{
dataPreperation(dat, uncertainty = "SE")
}
\arguments{
\item{dat}{Untransformed data table as shown in the example}

\item{uncertainty}{SE or SD}
}
\value{
An initialized landUse portfolio ready for initScenario
}
\description{
Preperation of the data for the robust optimization
}
