% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{defineConstraintCoefficients}
\alias{defineConstraintCoefficients}
\title{Internal function for the re-formulation of the long represantation of the constraints of the robust optimization into nLanduse colums (but still the original row length). Advantage is that the representation corresponds the expected form of the constraints of the lpSolveAPI package.}
\usage{
defineConstraintCoefficients(scenarioTable)
}
\arguments{
\item{scenarioTable}{The fully initialized scenario table.}
}
\value{
The coefficients of the objective function.
}
\description{
Internal function for the re-formulation of the long represantation of the constraints of the robust optimization into nLanduse colums (but still the original row length). Advantage is that the representation corresponds the expected form of the constraints of the lpSolveAPI package.
}
