% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{defineObjectiveCoefficients}
\alias{defineObjectiveCoefficients}
\title{Internal function for the re-formulation of the long represantation of the coefficients of the robust optimization into nLanduse Parameters. Advantage is that the shorter representation corresponds the expected form of the objective function of the lpSolveAPI package.}
\usage{
defineObjectiveCoefficients(scenarioTable)
}
\arguments{
\item{scenarioTable}{The fully initialized scenario table.}
}
\value{
The coefficients of the objective function.
}
\description{
Internal function for the re-formulation of the long represantation of the coefficients of the robust optimization into nLanduse Parameters. Advantage is that the shorter representation corresponds the expected form of the objective function of the lpSolveAPI package.
}
