% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDistanceToPerformanceScenario.R
\name{calcDistanceToPerformanceScenario}
\alias{calcDistanceToPerformanceScenario}
\title{Attach portfolio performance and distance to target}
\usage{
calcDistanceToPerformanceScenario(x)
}
\arguments{
\item{x}{An optimized optimLanduse object.}
}
\description{
The function calculates and attaches the portfolio performance and distance to target. See Gosling et al. Equations 10 and 11.
}
