% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSensitivity.R
\name{calcSensitivity}
\alias{calcSensitivity}
\title{Some heading}
\usage{
calcSensitivity(result, x = 0.01, digits = 3, fixDistance = NULL)
}
\arguments{
\item{x}{Dat}

\item{y}{Weiterer Parameter}
}
\description{
Text, short description
\code{\link{exampleData}} into to the expected format. The application of this function
is not mandatory
if you want to transform your data yourself or if your data is not formatted as
the example data. The application example on the asdf
\href{https://gitlab.gwdg.de/forest_economics_goettingen/optimlanduse}{GitLab project page}
provides information about the expected structure. Incomplete rows with NA-values are deleted and an error message is displayed.
}
\examples{
require(readxl)
dat <- read_xlsx(exampleData("exampleGosling_2020.xlsx"),
                 col_names = FALSE)
dat <- dataPreparation(dat, uncertainty = "SE", expVAL = "score")
}
