% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreparation.R
\name{dataPreparation}
\alias{dataPreparation}
\title{Preparation of the data for the robust optimization}
\usage{
dataPreparation(dat, uncertainty = "SE")
}
\arguments{
\item{dat}{Untransformed data table as shown in the example}

\item{uncertainty}{SE or SD}
}
\value{
An initialized landUse portfolio ready for initScenario
}
\description{
Preparation of the data for the robust optimization
}
