% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreparation.R
\name{dataPreparation}
\alias{dataPreparation}
\title{Preparation of the data for the robust optimization}
\usage{
dataPreparation(dat, uncertainty = "SE")
}
\arguments{
\item{dat}{Untransformed data table as shown in the example.}

\item{uncertainty}{Indicates whether the uncertainty shall be repesented by SE or SD. Please be aware that the respective chosen uncertainty must be included in the data. Best would be to consider the format of the exemplary data (database.xlsx) in the GitLab.}
}
\value{
An initialized landUse object ready for initScenario.
}
\description{
Preparation of the data for the robust optimization
}
