% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPreparation.R
\name{dataPreparation}
\alias{dataPreparation}
\title{Transform the data to the expected format}
\usage{
dataPreparation(dat, uncertainty = "SE", expVAL = "mean")
}
\arguments{
\item{dat}{Data frame formated as shown in the example (exemplary_data.xlsx).}

\item{uncertainty}{Indicates whether the uncertainty shall be repesented by SE or SD. Please be aware that the respective chosen uncertainty must be included in the data. Best would be to consider the format of the exemplary data (database.xlsx) in the GitLab.}

\item{expVAL}{Indicates how the expected Value should be represented.}
}
\value{
An initialized landUse object ready for initScenario.
}
\description{
The data must suit to the specific expected optimLanduse format. This function provides possibility to easily transform data in the form of the exemplary data LINK to the expected format. Application of this function is thus not mandatory if you want to transform your data manually or, if your dat are not formated as in the example data.
}
\examples{
require(xlsx)
}
