% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{exampleData}
\alias{exampleData}
\title{Import examplary data}
\usage{
exampleData(fileName = "database.xlsx")
}
\arguments{
\item{fileName}{Name of example file. See 'details' section for further explanation.}
}
\description{
optimLanduse comes bundled with exemplary land-use options and indicator values. The
files can also be found on your computer in the package folder `inst/extdata`. Aim is to
enable quick application of the package and to view the expected structure of the data.
}
\details{
\emph{database.xlsx} and \emph{databaseShrinked.xlsx} are exerpts grom Claff et al .... tbd. Volker
}
\examples{
require(readxl)
path <- exampleData()
read_xlsx(path, col_names = FALSE)
path <- exampleData("database.xlsx")
read_xlsx(path, col_names = FALSE)
path <- exampleData("databaseShrinked.xlsx")
read_xlsx(path, col_names = FALSE)
}
