% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{exampleData}
\alias{exampleData}
\title{Import examplary data}
\usage{
exampleData(fileName = NULL)
}
\arguments{
\item{fileName}{Name of example file. If `NULL`, a list of all example files will provided.}
}
\description{
optimLanduse comes bundled with exemplary land-use options and indicator values. The
files can also be found on your computer in the package folder `inst/extdata`. Aim is to
enable quick application of the package and to view the expected structure of the data.
}
\details{
This function is adopted from the example function of the
readxl pakage (1.3.1-2).
}
\examples{
exampleData()
exampleData("dataset.xlsx")
}
