% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{exampleData}
\alias{exampleData}
\title{Attach examplary data}
\usage{
exampleData(fileName = "database.xlsx")
}
\arguments{
\item{fileName}{Name of example file. See 'details' section for further explanation of all provided examples.}
}
\value{
The path to the example file on your computer.
}
\description{
optimLanduse comes bundled with exemplary data for land-use optimization. The
files can also be found on your computer in the package folder `./extdata`. These examples
provide some quick applications of the package for demonstration and an example of the expected structure
of the data. Consider also the
\href{https://gitlab.gwdg.de/forest_economics_goettingen/optimlanduse}{GitLab project page} for
exemplary applications of the package.
}
\details{
\emph{database.xlsx} and \emph{databaseShrinked.xlsx} are excerpts from Clough et al. 2016.
}
\examples{
require(readxl)
path <- exampleData()
read_xlsx(path, col_names = FALSE)
path <- exampleData("database.xlsx")
read_xlsx(path, col_names = FALSE)
path <- exampleData("databaseShrinked.xlsx")
read_xlsx(path, col_names = FALSE)

}
\references{
Clough, Y., Krishna, V., Corre, M. et al. Land-use choices follow profitability at the expense of
ecological functions in Indonesian smallholder landscapes. Nat Commun 7, 13137 (2016). https://doi.org/10.1038/ncomms13137
}
