% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initScenario.R
\name{initScenario}
\alias{initScenario}
\title{Initialize the robust optimization}
\usage{
initScenario(
  coefTable,
  uValue = 3,
  optimisticRule = "expectation",
  fixDistance = NULL
)
}
\arguments{
\item{coefTable}{Parameter and uncertainties in a specific format. Usage of the 'dataPreparation' function is recommended to make sure, the format requiremnts are met.}

\item{uValue}{u Value.}

\item{optimisticRule}{Either *expectation* or *uncertaintyAdjustedExpectation*. It indicates whether the optimistic outcomes of an indicator are directly reflected by the *expectation* or if the indicator is calculated *adjusted* by *expectation* + *uncertainty*.}

\item{fixDistance}{tbd.}
}
\value{
An initialized landUse object ready for optimization.
}
\description{
The function translates the prepared lanUse object into a solvable *optimLanduse* S3 object. Aim of the separation of the initialization and the optimization is to save time in the (possibly time-cunsuming) optimization.
}
