% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initScenario.R
\name{initScenario}
\alias{initScenario}
\title{Initialize the robust optimization}
\usage{
initScenario(coefTable, uValue = 3, optimisticRule = "expectation")
}
\arguments{
\item{coefTable}{See the exemplary import folder.}

\item{uValue}{u Value.}

\item{optimisticRule}{Either *expectation* or *uncertaintyAdjustedExpectation*. It indicates weather the optimistic outcomes of an indicator are directly reflected by the *expectation* or if the indicator is *adjusted*.}
}
\value{
An initialized landUse portfolio ready for optimization.
}
\description{
The function translates the indicators values and uncertainties for the land-use options into a solvable *optimLanduse* object.
}
