% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initScenario.R
\name{initScenario}
\alias{initScenario}
\title{Initialize the robust optimization}
\usage{
initScenario(
  coefTable,
  uValue = 3,
  optimisticRule = "expectation",
  fixDistance = NULL
)
}
\arguments{
\item{coefTable}{Coefficient table in the specific optimLanduse format.}

\item{uValue}{u Value.}

\item{optimisticRule}{Either \emph{expectation} or \emph{uncertaintyAdjustedExpectation}.
It indicates whether the optimistic outcomes of an indicator are directly
reflected by the expectation or if the indicator is adjusted by expectation +
uncertainty in case "more is better", expectation - uncertainty respectively when "less is better".}

\item{fixDistance}{tbd. Kai B.}
}
\value{
An initialized optimLanduse S3 object ready for optimization.
}
\description{
The function is used to initialize a \emph{optimLanduse} S3 object from the formated
coefficients table.
}
\details{
The expected format is explained in the example on
 \href{https://gitlab.gwdg.de/forest_economics_goettingen/optimlanduse}{GitLab}.
 Usage of \code{\link{dataPreparation}} is recommended to ensure that
 the format requirements are met.

 Aim of the separation of the initialization and the optimization is to save
 calculation time. The separated function calls allow the user to perform multiple
 optimization-runs from one initialized object. This could save time in batch
 applications.
}
\examples{
require(readxl)
dat <- read_xlsx(exampleData("databaseShrinked.xlsx"), col_names = FALSE)
dat <- dataPreparation(dat)
init <- initScenario(dat, uValue = 2, optimisticRule = "expectation", fixDistance = NULL)
}
