% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveScenario.R
\name{solveScenario}
\alias{solveScenario}
\title{Perform the optimization}
\usage{
solveScenario(x, digitsPrecision = 4, lowerBound = 0)
}
\arguments{
\item{x}{The initialized *optimLanduse* object.}

\item{digitsPrecision}{Precision.}
}
\value{
A solved landUse portfolio ready for export or further data processing.
}
\description{
The function solves the initialized *optimLanduse* object.
}
