% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveScenario.R
\name{solveScenario}
\alias{solveScenario}
\title{Perform the optimization}
\usage{
solveScenario(x, digitsPrecision = 4, lowerBound = 0, upperBound = 1)
}
\arguments{
\item{x}{The initialized *optimLanduse* object.}

\item{digitsPrecision}{Precision.}

\item{lowerBound}{An optional vector lower bounds for the land-use options. Must be in the dimension of the land-use options if delivered.}

\item{upperBound}{An optional vector upper bounds for the land-use options. Must be in the dimension of the land-use options if delivered.}
}
\value{
A solved landUse portfolio ready for export or further data processing.
}
\description{
The function solves the initialized *optimLanduse* object.
}
