% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveScenario.R
\name{solveScenario}
\alias{solveScenario}
\title{Perform the robust optimization}
\usage{
solveScenario(x, digitsPrecision = 4)
}
\arguments{
\item{x}{The *optimLanduse* object.}

\item{digitsPrecision}{Precision.}
}
\value{
A solved landUse portfolio.
}
\description{
The function solves the initialized *optimLanduse* object.
}
