% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveScenario.R
\name{solveScenario}
\alias{solveScenario}
\title{Perform the optimization}
\usage{
solveScenario(x, digitsPrecision = 4, lowerBound = 0, upperBound = 1)
}
\arguments{
\item{x}{The initialized \emph{optimLanduse} object.}

\item{digitsPrecision}{Precision of the solution.}

\item{lowerBound}{An optional vector lower bounds for the land-use options. Must be in the dimension of the land-use options if delivered.}

\item{upperBound}{An optional vector upper bounds for the land-use options. Must be in the dimension of the land-use options if delivered.}
}
\value{
A solved landUse portfolio ready for export or further data processing.
}
\description{
The function solves the initialized \emph{optimLanduse} object.
}
\details{
The foundations of the solver are explained in the paer of Knoke et al. (2016)
}
\examples{
require(readxl)
dat <- read_xlsx(exampleData("databaseShrinked.xlsx"),
                 col_names = FALSE)
dat <- dataPreparation(dat)
init <- initScenario(dat, uValue = 2,
                     optimisticRule = "expectation",
                     fixDistance = NULL)
result <- solveScenario(x = init)

}
\references{
Knoke, T., Paul, C., Hildebrandt, P. et al. (2020): Compositional diversity
of rehabilitated tropical lands supports multiple ecosystem services and
buffers uncertainties. \emph{Nat Commun} \strong{7}, 11877. \url{https://doi.org/10.1038/ncomms11877}
}
